% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupProject.R
\name{BackupProject}
\alias{BackupProject}
\title{Backup Project Directory}
\usage{
BackupProject(
  path_from = getwd(),
  subdir_to = "Projects",
  path_to = NULL,
  exclude = c(".DS_Store", ".Rproj.user", ".git"),
  recreate = TRUE
)
}
\arguments{
\item{path_from}{Path from where the folders should be copied (project
directory location).}

\item{subdir_to}{Subdirectory within shared drive to back up to}

\item{path_to}{Path to where the folders should be copied (P drive, only used
if specified).}

\item{exclude}{files/folders NOT to be backed up to the P-drive (useful for
larger files that don't change often). Currently not used.}

\item{recreate}{should backup be created from the ground up?
(can take longer, but useful for projects with many changes)}
}
\value{
This function has verbose output to ensure the back up is working, and
 ultimately returns a success indicator that's returned by file.copy.
}
\description{
This function backs up a CIDA project to the shared (P) CIDA drive. The backup directory
can either be existing (in which only changed files/folders are updated), or
nonexisting, in which case a full project backup is created.
}
