% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cida_word_document.R
\name{cida_word_document}
\alias{cida_word_document}
\title{A wrapper fn for CIDA word doc output}
\usage{
cida_word_document(...)
}
\arguments{
\item{...}{passed to rmarkdown::word_document}
}
\description{
This is a wrapper for the rmarkdown::word_document function.
}
