% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{CIDADefaults}
\alias{CIDADefaults}
\title{Creates a wrapper around CIDA defaults}
\usage{
CIDADefaults(
  .getter = function() NULL,
  .setter = function() NULL,
  .parent = NULL,
  analyst = NULL
)
}
\description{
Creates a wrapper around CIDA defaults
}
