% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{CIDADefaultsModel}
\alias{CIDADefaultsModel}
\title{The model for CIDA defaults.}
\usage{
CIDADefaultsModel(
  analyst = NULL,
  github_host = "github.com",
  github_protocol = "https",
  github_username = NULL,
  github_password = NULL
)
}
\description{
The model for CIDA defaults.
}
