% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{CIDAProject}
\alias{CIDAProject}
\title{Class to encapsulate a CIDA project.}
\usage{
CIDAProject(
  .path = character(0),
  project_name = NULL,
  principal_investigator = NULL,
  data_location = NULL,
  git_location = NULL,
  analyst = NULL
)
}
\arguments{
\item{project_name}{The name for the current project.}

\item{principal_investigator}{The principal investigator for the project.}

\item{data_location}{The location of the data for the project.}

\item{git_location}{The location of git remote for the project.}

\item{analyst}{The analyst for the project.}

\item{path}{The path to the project JSON file}
}
\description{
Class to encapsulate a CIDA project.
}
