% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{CIDAProjectModel}
\alias{CIDAProjectModel}
\title{Class to encapsulate a CIDA project.}
\usage{
CIDAProjectModel(
  project_name = NULL,
  principal_investigator = NULL,
  data_location = NULL,
  git_location = NULL,
  analyst = NULL,
  metadata = NULL
)
}
\arguments{
\item{project_name}{The name for the current project.}

\item{principal_investigator}{The principal investigator for the project.}

\item{data_location}{The location of the data for the project.}

\item{git_location}{The location of git remote for the project.}

\item{analyst}{The analyst for the project.}

\item{metadata}{User-specified metadata, can be used to customize templates
with extra information not stored in the provided fields.}
}
\description{
Class to encapsulate a CIDA project.
}
