% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{clear_project_undo_cache}
\alias{clear_project_undo_cache}
\title{copy local files to network path
prompt for updated remote files}
\usage{
clear_project_undo_cache()
}
\description{
copy local files to network path
prompt for updated remote files
}
