% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{convertIntervalNotation}
\alias{convertIntervalNotation}
\title{Convert Interval Notation}
\usage{
convertIntervalNotation(x)
}
\arguments{
\item{x}{a character vector to be converted}
}
\value{
a character vector of same length of x converted
}
\description{
Converts a vector from Interval Notation to less than equal to, less than,
etc.
}
\keyword{interval}
\keyword{notation}
