% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{create_github_repository}
\alias{create_github_repository}
\title{Creates a new GitHub repository. This function is intended for interactive
use, and wraps the more specific create_empty_github_repository() and 
create_github_repository_from_template() functions. If you do not require an
interactive component, use one of the more specific functions instead.}
\usage{
create_github_repository(
  name,
  description = NULL,
  visibility = "internal",
  template_name = NULL
)
}
\description{
Creates a new GitHub repository. This function is intended for interactive
use, and wraps the more specific create_empty_github_repository() and 
create_github_repository_from_template() functions. If you do not require an
interactive component, use one of the more specific functions instead.
}
