% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{ctcrw_interpolation}
\alias{ctcrw_interpolation}
\title{Interpolate GPS locations at a set time interval using a continuous time correlated
random walk (ctcrw) model}
\usage{
ctcrw_interpolation(
  data,
  innerBuff,
  returnBuff,
  duration,
  timestep,
  showPlots = F,
  theta = c(8, 2)
)
}
\arguments{
\item{data}{Tracking data downloaded directly from movebank using track2KBA::move2kba.}

\item{innerBuff}{minimum distance (km)  from the colony to be considered in a trip,
value passed to track2KBA::tripSplit()}

\item{returnBuff}{outer buffer (km)  to capture incomplete return trips to be considered in a trip,
value passed to track2KBA::tripSplit()}

\item{duration}{minimum time (hrs) away from colony  to be considered in a trip,
value passed to track2KBA::tripSplit()}

\item{timestep}{string indicating time step for track interpolation, eg. '10 min', '1 hour', '1 day'}

\item{showPlots}{TRUE/FALSE should plots of interpolated tracks against original data be produced}

\item{theta}{starting values for ctcrw parameter optimization, see ?crawl::crwMLE for details}
}
\description{
This function is a wrapper for momentuHMM::crawlWrap(), which uses the crawl
package to fit ctcrw model to GPS tracks at a user-defined time interval. The
function is currently designed to handle GPS data from centra place foraging birds.
It takes tracking data downloaded from Movebank using track2KBA::move2KBA().
Initial data processing identifies foraging trips away from the colony using
track2KBA::tripSplit(). The function returns a list with three SpatialPoints objects:
(1) original tracking data, (2) colony location, and (3) interpolated locations. All
spatial objects are in the same custom Lambert equal area projection centered on the
colony.
}
\examples{
dataset <- track2KBA::move2KBA(movebankID = 1895716931,
                              user = rstudioapi::askForPassword(prompt = 'Movebank username:'),
                              password = rstudioapi::askForPassword(prompt = 'Movebank password:'),
                              filename = NULL)

interp_data <- ctcrw_interpolation(data = dataset, innerBuff  = 5, returnBuff = 10,
                           duration   = 2,timestep = '60 min',show_plots = T)
}
