% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{current_project}
\alias{current_project}
\title{Retrieve the currently active CIDA project.}
\usage{
current_project(project_root = NULL)
}
\arguments{
\item{project_root}{The root directory for the project, or NULL. If the 
root directory is not specified, uses getwd() to obtain the current working
directory.}
}
\description{
This function will recursively search for a CIDA project config file, starting
with the current working directory and traversing upwards toward the root directory.
}
\cidatools{current_project}{project}
