% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_table.R
\name{desc_table}
\alias{desc_table}
\alias{desc_table.tbl_df}
\title{Create a descriptive table (table one)}
\usage{
desc_table(data, ...)

\method{desc_table}{tbl_df}(data, rowvars, colvar = NULL, freq = NULL,
  mean_sd = NULL, median_iqr = NULL, rowvar_names = NULL,
  incl_missing = TRUE, incl_pvalues = FALSE, unlist = TRUE,
  sort = TRUE, tight = TRUE, verbose = TRUE,
  allow_large_categories = FALSE, binary_first = FALSE,
  add_all = FALSE)
}
\arguments{
\item{data}{the tibble, data frame or design object of the data you which to
characterize}

\item{rowvars}{a vector of row positions or names. For tibbles see details.}

\item{colvar}{the position or name of the variable to stratify by, set
to \code{NULL} for no stratification.(p values cannot be returned if
\code{NULL}). for tibbles see details.}

\item{freq}{variables you want counts for. If \code{NULL} then set to all
charcter or factor variables in \code{rowvars}}

\item{mean_sd}{variables you want mean and standard deviations for.
If \code{NULL} then set to all
numeric (integer or double) variables in \code{rowvars}}

\item{median_iqr}{variables you want median, 25^th^ and 75^th^ percentiles
for. If \code{NULL} no variables will return a median and interquartile
range.}

\item{rowvar_names}{an optional vector of row names to be used for
variables. Maybe a named vector/list or an unnamed ordered vector.
If named not all rowvars must be present (see details).}

\item{incl_missing}{set to \code{TRUE} to include missing values (default)}

\item{incl_pvalues}{set to \code{TRUE} to include p values (p values are only
calculated on non missing observations) (default = \code{FALSE})}

\item{unlist}{should outputs be unlisted (default)}

\item{sort}{should variables be sorted by output}

\item{tight}{if \code{TRUE} (default) no spaces between numbers and
parenthesis}

\item{verbose}{should both levels of binary variables be printed.}

\item{allow_large_categories}{= FALSE,}

\item{binary_first}{should binary categoricals be placed at the top?
Ignored if `sort = FALSE`.}

\item{add_all}{= FALSE}
}
\value{
a tibble, dataframe or table depending on input.
}
\description{
This is a function created to provide characteristics of a study group with
an option to stratify by some variable (usually an exposure) The output of
this function is designed to be used with pander or kable in rmarkdown,
but all row
name formatting can be removed with the option: emphasis = 'n'.
}
\details{
If no stratification variable is provided, summary statistics on the
entire sample are provided. No p-values can be provided in this case.
If a design object is passed in lieu of a data frame, weighted numbers
using the survey package are provided.
(The survey package must be installed in this case.)

To use the lineBreaks use \code{panderOptions('keep.line.breaks', TRUE)}
}
\section{Methods (by class)}{
\itemize{
\item \code{tbl_df}: tibble descriptive table
}}

\keyword{characteristic}
\keyword{descriptive}
\keyword{table1}
\keyword{tableone}
