% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fns.R
\name{find_nearest}
\alias{find_nearest}
\alias{getlower}
\title{Find the nearest observation to another observation  - To be deprecated}
\usage{
find_nearest(
  x,
  y,
  direction = c("both", "ascending", "descending"),
  returnIndex = FALSE
)

getlower(x, y, upper = FALSE)
}
\arguments{
\item{x}{first value}

\item{y}{second value}

\item{direction}{default = both, ascending for only y matches before x and
descending for only y matches after x.}

\item{returnIndex}{should an index of the mathced y values be returned
instead of the matched list.}

\item{upper}{upper value?}
}
\value{
a list of length 2 with a y matched to every x, note if direction =
'ascending' or 'descending', NAs will be returned for x values with no y
values before or after, respectively. OR an index of matched y values if
\code{returnIndex = TRUE}.

indexes of y for each x
}
\description{
This function finds the nearest y to every x. Y's may be duplicated.
}
\section{Functions}{
\itemize{
\item \code{getlower()}: function for finding lower(upper) value

}}
\references{
This function borrowed heavily from this stack exchange post:
https://stats.stackexchange.com/questions/161379/quickly-finding-nearest-time-observation
}
\keyword{findnearest}
