% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{get_default_credentials}
\alias{get_default_credentials}
\title{Read default GitHub credentials from the CIDA defaults file.}
\usage{
get_default_credentials()
}
\description{
Read default GitHub credentials from the CIDA defaults file.
}
