% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{get_default_value}
\alias{get_default_value}
\title{Get a specific user level default}
\usage{
get_default_value(parameter)
}
\arguments{
\item{parameter}{The name of the specific parameter to lookup.}
}
\value{
The user level defaults named parameter
}
\description{
This function returns a specific user level defaults set in ~/cida_defaults.dcf
}
\keyword{options}
\keyword{path}
