% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{get_defaults}
\alias{get_defaults}
\title{Get a list of all user level defaults}
\usage{
get_defaults()
}
\value{
The user level defaults
}
\description{
This function returns the user level defaults set in ~/cida_defaults.dcf
}
\keyword{options}
\keyword{path}
