% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_defaults_path}
\alias{get_defaults_path}
\title{Gets the path to the defaults file.}
\usage{
get_defaults_path()
}
\description{
This is a function here because of R's compile-time evaluation of package-level
variables.
}
