% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{get_github_credentials}
\alias{get_github_credentials}
\title{Retrive a GitHub token to be used for CIDAtools functionality.}
\usage{
get_github_credentials()
}
\description{
This function will first attempt to retrieve a GitHub token from CIDADefaults, but will fall back to Git Credential Manager.
}
