% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{get_global_default_analyst}
\alias{get_global_default_analyst}
\title{Get Default Analyst Value}
\usage{
get_global_default_analyst()
}
\value{
The default analyst or an empty string if not set.
}
\description{
This function allows you to get the default analyst.
}
\keyword{Analyst}
\keyword{options}
