% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{get_global_default_path}
\alias{get_global_default_path}
\title{Get Default drive path}
\usage{
get_global_default_path()
}
\value{
The default path
}
\description{
This function returns the user level default drive path set in ~/cida_defaults.dcf
}
\keyword{options}
\keyword{path}
