% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{get_project_analyst}
\alias{get_project_analyst}
\title{Get Project Analyst}
\usage{
get_project_analyst()
}
\value{
A character string with the analyst name
}
\description{
This function returns the Project Analyst Name. If none exists, it
will return the value of CIDAtools.analyst option or blank if the option
is not set.
}
\keyword{Analyst}
\keyword{ProjData}
\keyword{options}
