% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{get_project_data}
\alias{get_project_data}
\title{Get data for project}
\usage{
get_project_data(param)
}
\arguments{
\item{param}{Project parameter to be gotten}
}
\description{
Allows you to get misc project data parameters
}
