% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{get_project_github}
\alias{get_project_github}
\title{Get Project GitHub Location}
\usage{
get_project_github()
}
\value{
A character string with the project GitHub
}
\description{
This function returns the Project GitHub location or blank if it's not set.
}
\keyword{ProjData}
\keyword{ProjectGitHub}
\keyword{options}
