% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{get_project_location}
\alias{get_project_location}
\title{Get Project data location on CIDA Drive}
\usage{
get_project_location(path = "")
}
\arguments{
\item{path}{(optional) a relative path to a particular place in the project}
}
\value{
full (absolute) file path including the project location on CIDA drive
}
\description{
Get Project data location on CIDA Drive
}
\examples{
# Read data from current project
\dontrun{
df <- read.csv(ProjectLocation("DataRaw/my_proj_data.csv"))
}

}
