% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{get_project_name}
\alias{get_project_name}
\title{Get Project Name}
\usage{
get_project_name()
}
\value{
A character string with the project name
}
\description{
This function returns the Project Name or blank if none exists.
}
\keyword{ProjData}
\keyword{ProjectName}
\keyword{options}
