% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{get_project_pi}
\alias{get_project_pi}
\title{Get PI Name}
\usage{
get_project_pi()
}
\value{
A character string with the PI name
}
\description{
This function returns the PI Name or blank if none exists.
}
\keyword{PI}
\keyword{ProjData}
\keyword{options}
