% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fns.R
\name{list_kables}
\alias{list_kables}
\title{List tables with the same columns - To be deprecated}
\usage{
list_kables(tabs, bo = c("striped", "condensed"), ...)
}
\arguments{
\item{tabs}{A list of data frames or tibbles with the same number and names of columns}

\item{bo}{boostrap options to pass to kableExtra}

\item{...}{other options to be passed to kable, such as caption, align, etc.}
}
\value{
HTML or Latex output for pretty tables
}
\description{
This function will print a group of tables together in a decently pretty way.
May need a bit of finagling.
}
