% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnRow.R
\name{MissingCont}
\alias{MissingCont}
\title{Internal functions for returning the number of missing for continuous
variable}
\usage{
MissingCont(x, y)
}
\arguments{
\item{x}{the row variable}

\item{y}{the column variable}
}
\description{
Internal functions for returning the number of missing for continuous
variable
}
