% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPExampleData.R
\docType{data}
\name{murres}
\alias{murres}
\title{Example GPS tracks from thick-billed murres breeding at Coats Island, NU}
\format{
GPS tracking data formatted to fit functions in track2KBA.
Data are a list object of length two, containing tracking data (accessed using: murres$data)
and study site location information (accessed using: murres$site).
}
\source{
Downloaded from the Arctic Animal Movement Archive project 'Thick-billed murre Elliott Coats 2010'
\url{https://www.movebank.org/cms/webapp?gwt_fragment=page=studies,path=study248994009} using
track2KBA::move2KBA()
}
\usage{
murres
}
\description{
Tracking data from thick-billed murres breeding at a colony on Coats Islands, NU, in 2010.
}
\keyword{datasets}
