% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{new_backup_project}
\alias{new_backup_project}
\title{copy local files to network path
prompt for updated remote files}
\usage{
new_backup_project(
  path_from = getwd(),
  path_to = NULL,
  exclude = c(".DS_Store", ".Rproj.user", ".git"),
  recreate = FALSE,
  data_only = TRUE,
  readme = TRUE
)
}
\description{
copy local files to network path
prompt for updated remote files
}
