% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{nLevelsP}
\alias{nLevelsP}
\title{Get pretty number of levels}
\usage{
nLevelsP(x)
}
\arguments{
\item{x}{factor}
}
\value{
Number of rows with big.mark = , and trim = T
}
\description{
Just a wrapper for format(nlevels) with big.mark = , and trim = T
}
\keyword{prettynlevels}
