% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{nrowP}
\alias{nrowP}
\title{Get pretty numbers of rows}
\usage{
nrowP(x)
}
\arguments{
\item{x}{data frame or matrix}
}
\value{
Number of rows with big.mark = , and trim = T
}
\description{
Retrieve the number of rows in dataframe of matrix with commas inserted for nice reports.
}
\keyword{prettynrow}
