% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{open_project}
\alias{open_project}
\title{Open Project
This function sets up the project to work on so the paths can easily be determined.
When both paths are specified path functions will reference the local copy.
When one is specified path functions will reference the local or remote copy whichever was specified.
Future updates will add some functionality to automate tasks.}
\usage{
open_project(local_project_folder = "", remote_project_folder = "")
}
\arguments{
\item{local_project_folder}{This is a local copy of the project folder}

\item{remote_project_folder}{This is the location of the shared drive copy of the project folder}
}
\description{
Open Project
This function sets up the project to work on so the paths can easily be determined.
When both paths are specified path functions will reference the local copy.
When one is specified path functions will reference the local or remote copy whichever was specified.
Future updates will add some functionality to automate tasks.
}
