% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp2KBA}
\alias{opp2KBA}
\title{Converts movebank data to format required for track2KBA}
\usage{
opp2KBA(data)
}
\arguments{
\item{data}{A dataframe obtained using OPPTools::opp_download_data}
}
\value{
Returns a list object of length two, containing tracking data
(accessed using: dataset$data) and study site location information
(accessed using: dataset$site).
}
\description{
Takes tracking data downloaded from Movebank using OPPTools::opp_download_data
and converts it to the format needed for track2KBA
}
\details{
This extracts location, timestamp and deployment data from movebank data
and returns a list of dataframes that can be passed to functions in track2KBA.
This is useful if the user wants to filter the data downloaded from movebank based on
fields contained within the reference data (e.g. sex, animal_reproductive_condition)
}
\examples{
my_data <- opp_download_data(study = c(1247096889),login = NULL, start_month = NULL,
                            end_month = NULL,season = NULL)

my_track2kba <- opp2KBA(data = my_data)

}
