% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_download_data}
\alias{opp_download_data}
\title{Download OPP tracking data from Movebank}
\usage{
opp_download_data(
  study,
  login = NULL,
  start_month = NULL,
  end_month = NULL,
  season = NULL
)
}
\arguments{
\item{study}{List of Movebank project ids.}

\item{login}{Stored Movebank login credentials if provided, otherwise function
will prompt users to enter credentials.}

\item{start_month}{Earliest month (1-12) to include in output.}

\item{end_month}{Latest month (1-12) to include in output.}

\item{season}{Vector describing the season data can be applied to, eg. 'Breeding (Jun-Jul)'}
}
\description{
This function downloads OPP tracking data from Movebank and returns a
dataframe with combined tracking and reference data for all deployments.
}
\details{
The function can be passed a list of movebank study IDs and will append
data from all studies.
}
\examples{

# download ANMU project data from two studies, for May only
my_data <- opp_download_data(study = c(1895716931, 1897273090),
login = NULL, start_month = 5, end_month = 5, season = 'Incubation')

}
