% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_explore_trips}
\alias{opp_explore_trips}
\title{Explore trip data within given tracks}
\usage{
opp_explore_trips(data)
}
\arguments{
\item{data}{Movebank data as returned by opp_download_data.}
}
\description{
This function calculates the distance from the
study site for each GPS point within a Movebank object and
then produces track time vs. distance from origin site plots.
Using this function will allow you to assign reasonable
estimates for minimum and maximum trip duration and distance
for the opp_get_trips function.
}
\examples{
data(murres)
opp_explore_trips(murres)

}
