% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_get_trips}
\alias{opp_get_trips}
\title{Identify foraging trips in tracking data}
\usage{
opp_get_trips(
  data,
  innerBuff = 5,
  returnBuff = 20,
  duration = 2,
  gapLimit = 100,
  missingLocs = 0.2,
  showPlots = TRUE
)
}
\arguments{
\item{data}{Tracking data formated using track2KBA or opp2KBA}

\item{innerBuff}{Minimum distance (km) from the colony to be in a trip.
Used to label trips as 'Non-trip'. Defaults to 5}

\item{returnBuff}{Outer distance (km) to capture trips that start and end
away from the colony. Used to label trips as 'Incomplete'. Defaults to 20.}

\item{duration}{Minimum trip duration (hrs)}

\item{gapLimit}{Maximum time between points to be considered too large to be
a contiguous tracking event. Can be used to ensure that deployments on the
same animal in different years do not get combined into extra long trips.
Defaults to 100 days.}

\item{missingLocs}{Proportion (0-1) of trip duration that a gap in consecutive
locations should not exceed. Used to label trips as 'Gappy'. Defaults to 0.2.}

\item{showPlots}{Logical (T/F), should plots showing trip classification by generated?}
}
\description{
Uses criteria related to distance from colony, trip duration, and size of gaps
in tracking data to identify and classify trips from a nest or colony. It is
a wrapper for track2KBA::tripSplit that applies custom criteria for classifying
trips.
}
\examples{
my_data <- opp_download_data(study = c(1247096889),login = NULL, start_month = NULL,
                            end_month = NULL,season = NULL)

my_track2kba <- opp2KBA(data = my_data)

my_trips <- opp_get_trips(data = my_track2kba, innerBuff  = 5, returnBuff = 20,
                         duration  = 2, gapLimit = 100, missingLocs = 0.2,
                         showPlots = TRUE)
}
