% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{opp_map}
\alias{opp_map}
\title{Plot raw tracks from Movebank download}
\usage{
opp_map(data, interactive = FALSE)
}
\arguments{
\item{data}{Movebank data as returned by opp_download_data.}

\item{interactive}{Logical (T/F), do you want to explore tracks with an interative map? Default FALSE.}
}
\description{
Quickly plot Movebank data downloaded
using opp_download_data to visualize tracks.
}
\examples{
data(murres)
opp_map(murres)

}
