% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{prep_ecotone}
\alias{prep_ecotone}
\title{Prepare raw Ecotone data for Movebank upload.}
\usage{
prep_ecotone(data, colony_lon, colony_lat, tz = "UTC")
}
\arguments{
\item{data}{Input Ecotone data to be modified.}

\item{colony_lon}{Longitude of home colony of tagged bird.}

\item{colony_lat}{Latitude of home colony of tagged bird.}

\item{tz}{Timezone of GPS timestamps. Default "UTC".}
}
\description{
This function modifies raw Ecotone GPS data to remove
any records without lat/long values, inserts a "behavior"
column to indicate when a tagged bird is at the colony,
adds a timestamp column ("Date_2") if not already there,
and removes any duplicate detections. The function also
inserts lat/long coordinates for the colony location for
periods when the bird is at the colony.
}
