% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{prep_pathtrack}
\alias{prep_pathtrack}
\title{Prepare Pathtrack data for Movebank upload.}
\usage{
prep_pathtrack(data)
}
\arguments{
\item{data}{Input Pathtrack data to be modified.}
}
\description{
This simple function processes Pathtrack data that has
been exported from Pathtrack Host software for Movebank
upload. It removes any records in Pathtrack data that
have `null` latitude or longitude values.
Unlike `prep_ecotone()`, this function makes no assumptions
on the bird's location when latitude/longitude are null.
}
