% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fns.R
\name{proj_location_handler}
\alias{proj_location_handler}
\title{helper function to cleanup project location}
\usage{
proj_location_handler(loc = "")
}
\description{
helper function to cleanup project location
}
