% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{ProjectAnalyst}
\alias{ProjectAnalyst}
\title{Get Project Analyst}
\usage{
ProjectAnalyst()
}
\value{
A character string with the analyst name
}
\description{
This function returns the Project Analyst Name. If none exists, it
will return the value of CIDAtools.analyst option or blank if the option
is not set.
}
