% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{ProjectLocation}
\alias{ProjectLocation}
\title{Get Project data location on CIDA Drive}
\usage{
ProjectLocation()
}
\value{
full (absolute) file path including the project location on CIDA drive
}
\description{
Get Project data location on CIDA Drive
}
\examples{
# Read data from current project
\dontrun{
df <- read.csv(ProjectLocation("DataRaw/my_proj_data.csv"))
}

}
