% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{read_model}
\alias{read_model}
\title{Function to load a project.json file into a CIDAProject object.}
\usage{
read_model(path, model)
}
\arguments{
\item{path}{The path to the project JSON file.}
}
\description{
This is an internal function which shouldn't be called by users directly.
}
