% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{read_raw_metadata}
\alias{read_raw_metadata}
\title{Function to load a project.json file into a CIDAProject object.}
\usage{
read_raw_metadata(path)
}
\arguments{
\item{path}{The path to the project JSON file.}
}
\description{
This is an internal function which shouldn't be called by users directly.
}
