% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{remove_global_default_analyst}
\alias{remove_global_default_analyst}
\title{Remove Default Analyst from ~/cida_defaults.dcf}
\usage{
remove_global_default_analyst()
}
\arguments{
\item{quiet}{should a message indicating result be returned, if TRUE will only
return TRUE or FALSE}
}
\value{
Message indicating sucess or failue
}
\description{
This function removes the default analyst set with set_global_default_analyst()
 from the users ~/cida_defaults.dcf.
}
\keyword{Analyst}
\keyword{remove}
