% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{remove_global_default_path}
\alias{remove_global_default_path}
\title{removes the default drive path in ~/cida_defaults.dcf}
\usage{
remove_global_default_path()
}
\description{
This function removed the default path at the user level to access the project
drive.
}
