% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{removeAnalyst}
\alias{removeAnalyst}
\title{Remove Default Analyst from ~/cida_defaults.dcf}
\usage{
removeAnalyst()
}
\value{
Message indicating success or failure
}
\description{
This function removes the default analyst set with set_global_default_analyst()
 from the users ~/cida_defaults.dcf.
}
