% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render}
\alias{render}
\title{Render an Rmarkdown document and write the output to the specified directory}
\usage{
render(output_dir = NULL, prefix = NULL, date_fmt = NULL)
}
\arguments{
\item{output_dir}{The directory where the output should be saved.}

\item{prefix}{The prefix of output file.}

\item{date_fmt}{A format string (e.g. '%Y%m%d-%H%M%S') used to format a
date string that is appended to the filename. If NULL, don't append date.}
}
\description{
Render an Rmarkdown document and write the output to the specified directory
}
