% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{retrieve_metadata_value}
\alias{retrieve_metadata_value}
\title{Function to retrieve a metadata value from file.}
\usage{
retrieve_metadata_value(self, name)
}
\description{
Function to retrieve a metadata value from file.
}
