% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{Round}
\alias{Round}
\title{Round and don't drop trailing zeros}
\usage{
Round(x, n)
}
\arguments{
\item{x}{numeric to be formatted}

\item{n}{number of digits for nsmall}
}
\value{
a character vector of same length of x converted
}
\description{
Shorter wrapper for format(x, digits = n, nsmall = n)
}
\details{
should not be used unless digits after a decimal are needed.
Note for numbers with leading zeros (ie. 0.0349) you will get one more
decimal place than n. (ie. \code{Round(O.0349, 2)} will return
\code{0.035})
}
\keyword{interval}
\keyword{notation}
