% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{round_pretty}
\alias{round_pretty}
\title{Round and don't drop trailing zeros}
\usage{
round_pretty(x, digits = 0, type = c("s", "d"))
}
\arguments{
\item{x}{numeric to be formatted}

\item{digits}{number of digits}

\item{type}{is n signifigant digits \code{s} or digits \code{d} after
the decimal}
}
\value{
a character vector of same length of x converted
}
\description{
Round and don't drop trailing zeros
}
\details{

}
\keyword{round}
