% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_tags.R
\name{roxy_tag_parse.roxy_tag_cidatools}
\alias{roxy_tag_parse.roxy_tag_cidatools}
\title{Parse the cidatools Roxygen tag.}
\usage{
\method{roxy_tag_parse}{roxy_tag_cidatools}(x)
}
\description{
This function is responsible for parsing the '@cidatools' symbol
used to document public CIDAtools functions. We expect that the cidatools tag 
should be followed by two words (i.e. '@cidatools <tag> <category>'). 
The first word denotes the tag, or name of the function as it appears in the 
documentation. The second word is the category of the function. The tag and
category allow us to link and categorize each R CIDAtools function with the 
equivalent implementations from the Python and CLI versions of CIDAtools.
}
