% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{rsync_project}
\alias{rsync_project}
\title{Rsync type backup of project updated files?
files copied in both directions flag/skip conflicting files?}
\usage{
rsync_project()
}
\description{
Rsync type backup of project updated files?
files copied in both directions flag/skip conflicting files?
}
