% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{set_default_value}
\alias{set_default_value}
\title{Sets a default parameter in ~/cida_defaults.dcf}
\usage{
set_default_value(parameter, value)
}
\arguments{
\item{parameter}{parameter name}

\item{value}{new value}
}
\description{
This function sets a default parameter at the user level.
}
