% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{set_full_project_path}
\alias{set_full_project_path}
\title{Sets the default full path to the project.}
\usage{
set_full_project_path(path = "")
}
\arguments{
\item{path}{full path to the project folder}
}
\value{
message indicating the path has been saved.
}
\description{
Sets the default full path to the project.
}
