% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{set_global_default_path}
\alias{set_global_default_path}
\title{Sets a default drive path in ~/cida_defaults.dcf}
\usage{
set_global_default_path(path = "")
}
\arguments{
\item{path}{Path to the main project(CIDA) drive}
}
\value{
Message that path was set.
}
\description{
This function sets a default path at the user level to access the project
drive. If the drive is not detected automatically this file will be used to
set a default path if set.
}
