% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{set_metadata_value}
\alias{set_metadata_value}
\title{Function to set a metadata file in the file.}
\usage{
set_metadata_value(self, name, value)
}
\description{
Function to set a metadata file in the file.
}
