% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{set_project_analyst}
\alias{set_project_analyst}
\title{Set Project Analyst}
\usage{
set_project_analyst(analyst_name)
}
\arguments{
\item{analyst_name}{A string containing the analyst name}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the  project analyst.
This will overwrite the current value if exists.
}
\keyword{Analyst}
\keyword{ProjData}
\keyword{options}
