% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{set_project_github}
\alias{set_project_github}
\title{Set Project GitHub Location}
\usage{
set_project_github(git_url = "")
}
\arguments{
\item{git_url}{A string containing the URL to the GitHub repository for this project.}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the Project's GitHub location.
This will overwrite the current value if exists.
}
\keyword{ProjData}
\keyword{location}
\keyword{options}
