% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{set_project_meta_data}
\alias{set_project_meta_data}
\title{Set data for project}
\usage{
set_project_meta_data(parameter, value)
}
\arguments{
\item{parameter}{Project Parameter to be set}

\item{value}{Value to set to project parameter}
}
\description{
Allows you to set misc project data parameters
for Project Name, Analyst, or PI recommend you use specific function
}
