% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{set_project_name}
\alias{set_project_name}
\title{Set Project Name}
\usage{
set_project_name(project_name)
}
\arguments{
\item{project_name}{A string containing the analyst name}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the  project name. This will overwrite the
current value if exists.
}
\keyword{ProjData}
\keyword{ProjectName}
\keyword{options}
