% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{set_project_pi}
\alias{set_project_pi}
\title{Set PI Name}
\usage{
set_project_pi(pi)
}
\arguments{
\item{pi}{A string containing the analyst name}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the Project's PI. This will overwrite the
current value if exists.
}
\keyword{PI}
\keyword{ProjData}
\keyword{options}
