% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{setPermanentAnalyst}
\alias{setPermanentAnalyst}
\title{Set Global Default Analyst Value}
\usage{
setPermanentAnalyst(analyst_name, update_template = FALSE)
}
\arguments{
\item{analyst_name}{A string containing the analyst name}

\item{update_template}{Bool whether or not to update analyst in the project template.}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the default analyst at the user level and
if possible changes the default analyst in the template for new projects.
}
