% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{SetProjectData}
\alias{SetProjectData}
\title{Set data for project}
\usage{
SetProjectData(parameter, value)
}
\arguments{
\item{parameter}{Project Parameter to be set}

\item{value}{Value to set to project parameter}
}
\description{
Allows you to set misc project data parameters
for Project Name, Analyst, or PI recommend you use specific function
}
