% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{SetProjectLocation}
\alias{SetProjectLocation}
\title{Set Project Location}
\usage{
SetProjectLocation(path)
}
\arguments{
\item{path}{A string containing the file path to the project location on CIDA drive}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the Project's location on the CIDA drive.
This will overwrite the current value if exists.
}
