% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{SetProjectName}
\alias{SetProjectName}
\title{Set Project Name}
\usage{
SetProjectName(ProjectName)
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the  project name. This will overwrite the
current value if exists.
}
