% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{setter_with_warn}
\alias{setter_with_warn}
\title{Function to act as a setter for CIDAProject properties.}
\usage{
setter_with_warn(self, name, value)
}
\description{
Function to act as a setter for CIDAProject properties.
}
