% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summ_fxns}
\alias{summ_fxns}
\alias{count_fxn}
\alias{mean_sd_fxn}
\alias{median_iqr_fxn}
\title{Summary Functions}
\usage{
count_fxn(data, count_var)

mean_sd_fxn(data, mean_sd_var)

median_iqr_fxn(data, median_var)
}
\arguments{
\item{data}{A tbl.}

\item{count_var}{A variable for which you want counts. If more than one
variable is needed use `purrr::map`.}

\item{mean_sd_var}{A variable for which you want a mean and standard
deviation. If more than one
variable is needed use `purrr::map`.}

\item{median_iqr_var}{A variable for which you want a median, and 25^th^ and
75^th^
percentile. If more than one
variable is needed use `purrr::map`.}
}
\description{
These are functions that provide count, mean and sd
and median and IQR summaries for tibbles.
They are wrappers for summarise functions that are designed to used
with `purrr::map` and `purrr::map_dfr` to create summaries of many
variables in long table form.
}
