% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{vec_power}
\alias{vec_power}
\title{Vectorized power estimates}
\usage{
vec_power(fun = stats::power.t.test, ...)
}
\arguments{
\item{fun}{What is the function to calculate power}

\item{...}{other arguments to pass to power_fn, possibly vectorized}
}
\value{
tibble of results
}
\description{
This function allows you to use power.t.test, power.prop.test, etc in
vectorized fashion and return a table of results
}
