% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{write_config}
\alias{write_config}
\title{Writes a CIDAProjectModel object to JSON}
\usage{
write_config(project_root, metadata, overwrite = FALSE)
}
\description{
Writes a CIDAProjectModel object to JSON
}
