% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{write_gitignore}
\alias{write_gitignore}
\title{Creates a .gitignore file at the given path if none exists.}
\usage{
write_gitignore(project_root)
}
\description{
Creates a .gitignore file at the given path if none exists.
}
