% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{write_model}
\alias{write_model}
\title{Function to write a model object to JSON.}
\usage{
write_model(path, instance)
}
\arguments{
\item{path}{Path to the metadata file to read/write.}

\item{instance}{The instance of the model object to write.}
}
\description{
Function to write a model object to JSON.
}
