% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{write_raw_metadata}
\alias{write_raw_metadata}
\title{Function to write a CIDAProject object to JSON.}
\usage{
write_raw_metadata(path, metadata)
}
\arguments{
\item{path}{Path to the metadata file to read/write.}

\item{metadata}{The CIDAProject object to write to file.}
}
\description{
Function to write a CIDAProject object to JSON.
}
