% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{write_rproj}
\alias{write_rproj}
\title{Creates an .Rproj file in the current directory if none exists.}
\usage{
write_rproj(project_root, metadata)
}
\description{
Creates an .Rproj file in the current directory if none exists.
}
