% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{write_templated_readme}
\alias{write_templated_readme}
\title{Writes a templated README to a project folder}
\usage{
write_templated_readme(
  project_root,
  template_name,
  metadata,
  overwrite = FALSE,
  subdir = NULL
)
}
\description{
Writes a templated README to a project folder
}
