% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{cache_helper}
\alias{cache_helper}
\title{Function to help with caching of the metadata file, to avoid unecessary
reads and writes.}
\usage{
cache_helper(path)
}
\arguments{
\item{path}{The path to the metadata file.}
}
\description{
Function to help with caching of the metadata file, to avoid unecessary
reads and writes.
}
