% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cida_drive_path.R, R/misc_fns.R
\name{CIDA_drive_path}
\alias{CIDA_drive_path}
\title{Get CIDA drive path}
\usage{
CIDA_drive_path(path = "")

CIDA_drive_path(path = "")
}
\arguments{
\item{path}{(optional) a path to a particular place in the CIDA drive}

\item{file}{(optional) Path to subdirectory/file within CIDA drive}
}
\value{
Full (absolute) file path of CIDA drive

full (absolute) file path of CIDA drive
}
\description{
This function attempts to get the proper path for the CIDA drive either on
Windows or Mac.

This function attempts to get the proper path for the CIDA drive either on Windows or Mac.
}
\examples{
# Read data from P1234PIname project
\dontrun{
df <- read.csv(CIDA_drive_path("Projects/P1234PIname/DataRaw/data.csv"))
}

# Read data from P1234PIname project
\dontrun{
df <- read.csv(CIDA_drive_path("Projects/P1234PIname/DataRaw/data.csv"))
}

}
