% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDA_word_document.R
\name{CIDA_word_document}
\alias{CIDA_word_document}
\title{Custom CIDA Word template}
\usage{
CIDA_word_document(...)
}
\arguments{
\item{...}{additional arguments provided to \@code{word_document}}
}
\description{
Loads additional template file
}
