% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPFunctions.R
\name{colCRS}
\alias{colCRS}
\title{Define a custom equal-area CRS centered on your study site}
\usage{
colCRS(data)
}
\arguments{
\item{data}{Movebank data as returned by opp_download_data.}
}
\description{
This function takes a Movebank data object and
creates an equal-area projection ceneterd on the Movebank
study site. In the case of central-place foraging seabirds,
this effectively equates to a CRS centered on the seabird
colony. The function returns a proj4 string.
}
\examples{
data(murres)
colCRS(murres)

}
