% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create Project Directory + readme files}
\usage{
create_project(
  project_root = NULL,
  project_name = NULL,
  principal_investigator = NULL,
  analyst = NULL,
  data_location = NULL,
  git_location = NULL,
  folders_to_create = NULL
)
}
\arguments{
\item{project_name}{Name of project, (required)}

\item{analyst}{Name of Analyst(s), (required)}

\item{data_location}{Location of project on CIDA Drive, or "" for blank}

\item{git_location}{Location project on GitHub}

\item{folders_to_create}{A list of the project subdirectories to create.
If NULL, uses the value of 'CIDA_PROJECT_DEFAULT_FOLDERS'.}

\item{path}{Where should they be created? Default is the working directory.}

\item{template}{Which subdirectories to create}

\item{pi}{Name of PI and credentials, or "" for blank}
}
\value{
Function to create a CIDA project locally. This will create the 
 CIDA project structure with subdirectories, .gitignore, READMEs, etc, but
 will not replace existing files.
}
\description{
This function creates the standard project organization structure for CIDA
within a folder that already exists.
}
\keyword{create_project}
\keyword{project}
\cidatools{create_project}{project}
