% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create Project Directory + readme files}
\usage{
create_project(
  path = getwd(),
  template = c("Admin", "Background", "Code", "DataRaw", "DataProcessed",
    "Dissemination", "Reports"),
  project_name,
  pi = "",
  analyst,
  data_location = "",
  git_location = ""
)
}
\arguments{
\item{path}{Where should they be created? Default is the working directory.}

\item{template}{Which subdirectories to create}

\item{project_name}{Name of project, (required)}

\item{pi}{Name of PI and credentials, or "" for blank}

\item{analyst}{Name of Analyst(s), (required)}

\item{data_location}{Location of project on CIDA Drive, or "" for blank}

\item{git_location}{Location project on GitHub}
}
\value{
This function creates the desired project subdirectories and readmes,
 as well as a standard .gitignore file files. It will not overwrite the file
 however if it does not exist. It does not return anything.
}
\description{
This function creates the standard project organization structure for CIDA
within a folder that already exists.
}
\seealso{
proj_setup() is the internal wrapper for this that gets called when
 using the RStudio GUI to create a project
}
\keyword{createproject}
\keyword{project}
