% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProject.R
\name{CreateProject}
\alias{CreateProject}
\title{Create Project Directory + readme files}
\usage{
CreateProject(
  template = c("01_Raw_data", "02_Analysis", "03_Working_files", "04_Processed",
    "05_Admin"),
  path = getwd(),
  ProjectName = "",
  PI = "",
  analyst = ""
)
}
\arguments{
\item{template}{Which subdirectories to create}

\item{path}{Where should they be created? Default is the working directory.}

\item{ProjectName}{Name of project, which should be the year of the dataset; "" for blank}

\item{PI}{Name and credentials of PI who provided raw data, i.e., lead contact for data questions; "" for blank}

\item{analyst}{Name of Analyst(s) who wrote the analysis code, i.e., lead contact for code questions; "" for blank}
}
\value{
This function creates the desired project subdirectories and readmes,
 as well as a standard .gitignore file files. It will not overwrite the file
 however if it does not exist. It does not return anything.
}
\description{
This function creates the standard project organization structure for CIDA
within a folder that already exists.
}
\keyword{createproject}
\keyword{project}
