% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{CreateProject}
\alias{CreateProject}
\title{Create Project Directory + readme files}
\usage{
CreateProject(
  path = getwd(),
  template = c("Admin", "Background", "Code", "DataRaw", "DataProcessed",
    "Dissemination", "Reports"),
  ProjectName = "",
  PI = "",
  analyst = "",
  datalocation = ""
)
}
\arguments{
\item{path}{Where should they be created? Default is the working directory.}

\item{template}{Which subdirectories to create}

\item{ProjectName}{Name of project, or "" for blank}

\item{PI}{Name of PI and credentials, or "" for blank}

\item{analyst}{Name of Analyst(s), or "" for blank}

\item{datalocation}{Location of project on CIDA Drive, or "" for blank}
}
\value{
This function creates the desired project subdirectories and readmes,
 as well as a standard .gitignore file files. It will not overwrite the file
 however if it does not exist. It does not return anything.
}
\description{
This function creates the standard project organization structure for CIDA
within a folder that already exists.
}
\seealso{
proj_setup() is the internal wrapper for this that gets called when
 using the RStudio GUI to create a project
}
\keyword{createproject}
\keyword{project}
