% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReadMe.R
\name{CreateReadMe}
\alias{CreateReadMe}
\title{Create ReadMe.md files}
\usage{
CreateReadMe(
  template = c("01_Raw_data", "02_Analysis", "03_Working_files", "04_Processed",
    "05_Admin"),
  path = getwd()
)
}
\arguments{
\item{template}{Which ReadMe.md files should be created? Default is all. Partial
matching does work.}

\item{path}{Where should they be created? Default is the working directory.}
}
\value{
This function creates the desired readme files. It will not overwrite the file
however if it does not exists. It does not return anything.
}
\description{
This function creates ReadMe.md in the CIDA templates for the standard
file structure
}
\keyword{ReadMe}
\keyword{ReadMe.md}
