% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{get_full_project_path}
\alias{get_full_project_path}
\title{Gets the currently set full path to the project}
\usage{
get_full_project_path()
}
\value{
full path to project
}
\description{
Gets the currently set full path to the project
}
