% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_project_drive_path}
\alias{get_project_drive_path}
\title{Get Project drive path}
\usage{
get_project_drive_path(file = "")
}
\arguments{
\item{file}{(optional) Path to subdirectory/file within the main project(CIDA) drive}
}
\value{
Full (absolute) file path of project(CIDA) drive
}
\description{
This function attempts to get the proper path for the Project(CIDA) drive either on
Windows or Mac automatically.  If the expected path is not found it tried to
load the project metadata path and if that fails it looks for a global default path
in the user cida_defaults.dcf file.
}
\examples{
# Read data from P1234PIname project
\dontrun{
df <- read.csv(get_project_drive_path("BRANCHES/Pulmonary/P1234PIname/DataRaw/data.csv"))
}

}
