% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{getProjectData}
\alias{getProjectData}
\title{Get data for project}
\usage{
getProjectData(param)
}
\arguments{
\item{param}{Project parameter to be gotten}
}
\description{
Allows you to get misc project data parameters
}
