% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{ProjectAnalyst}
\alias{ProjectAnalyst}
\title{Get Project Analyst}
\usage{
ProjectAnalyst()
}
\value{
A character string with the analyst name
}
\description{
This function returns the Project Analyst Name. If none exists, it
will return the value of CIDAtools.analyst option or blank if the option is not set.
}
\keyword{Analyst}
\keyword{ProjData}
\keyword{options}
