% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{ProjectLocation}
\alias{ProjectLocation}
\title{Get Project location on CIDA Drive}
\usage{
ProjectLocation(path = "")
}
\arguments{
\item{path}{(optional) a relative path to a particular place in the project}
}
\value{
full (absolute) file path including the project location on CIDA drive
}
\description{
Get Project location on CIDA Drive
}
\examples{
# Read data from current project
\dontrun{
df <- read.csv(ProjectLocation("DataRaw/my_proj_data.csv"))
}

}
