% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{ProjectName}
\alias{ProjectName}
\title{Get Project Name}
\usage{
ProjectName()
}
\value{
A character string with the project name
}
\description{
This function returns the Project Name or blank if none exists.
}
\keyword{ProjData}
\keyword{ProjectName}
\keyword{options}
