% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{ProjectPI}
\alias{ProjectPI}
\title{Get PI Name}
\usage{
ProjectPI()
}
\value{
A character string with the PI name
}
\description{
This function returns the PI Name or blank if none exists.
}
\keyword{PI}
\keyword{ProjData}
\keyword{options}
