% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_fns.R
\name{pvalr}
\alias{pvalr}
\title{Pretty p-values}
\usage{
pvalr(pvals, sig.limit = 0.001, digits = 3, html = FALSE, equal_sign = "")
}
\arguments{
\item{pvals}{a vector of numeric p-values}

\item{sig.limit}{A lower threshold below which to print "<sig.limit"}

\item{digits}{how many digits to print?}

\item{html}{uses the HTML symbol instead of normal < symbol}

\item{equal_sign}{character value to append to front of p-value}
}
\value{
A character vector of "pretty" p values
}
\description{
This function helps print p-values in RMD output
}
