% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_pvalue.R
\name{pvalue}
\alias{pvalue}
\title{Internal function for p-value calculation in cida_table1}
\usage{
pvalue(x, name, ...)
}
\arguments{
\item{x}{the row_variable}

\item{name}{the variables name}

\item{...}{arguments passed to S3 methods}
}
\description{
Internal function for p-value calculation in cida_table1
}
