% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizer_and_blinder_app.R
\name{randblinder_shiny_tool}
\alias{randblinder_shiny_tool}
\title{Randomizer and blinder tool}
\usage{
randblinder_shiny_tool()
}
\description{
See the inst/ folder for the main code for this function
}
