% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelWithColor.R
\name{read.xlsx.withcolor}
\alias{read.xlsx.withcolor}
\title{Read in xlsx with fill colour}
\usage{
read.xlsx.withcolor(file, colorColumns, sheet = NULL, header = T)
}
\arguments{
\item{file}{the path to the file you intend to read. Can be an xls or xlsx format.}

\item{colorColumns}{a vector of column numbers for which you want to read the colour}

\item{sheet}{NULL(default) for all sheets otherwise a vector of sheet numbers
or names to read.}

\item{header}{should the 1st row be read in as a header? defaults to T.}
}
\value{
A data frame for one sheet or a list of data frames for multiple sheets
}
\description{
Reads in the fill colour of excel workbooks. Creates a data frame for each
sheet in a list if mutliple sheets are requested. Creates a colour column for
each colour column specified.
}
\keyword{Excel}
\keyword{color}
\keyword{colour}
\keyword{xlsx}
