% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{removeAnalyst}
\alias{removeAnalyst}
\title{Remove Default Analyst from ~/.Rprofile}
\usage{
removeAnalyst(quiet = F)
}
\arguments{
\item{quiet}{should a message indicating result be returned, if TRUE will only
return TRUE or FALSE}
}
\value{
Message indicating sucess or failue
}
\description{
This function removes the default analyst set with setAnalyst() from the users
.Rprofile. If this is the only entry in .Rprofile it will remove the file as well.
}
\keyword{Analyst}
\keyword{remove}
