% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnRow.R
\name{returnRow}
\alias{returnRow}
\alias{returnRow.factor}
\alias{returnRow.numeric}
\alias{returnRow.character}
\alias{returnRow.MedIQR}
\title{These are internal functions for returning the correct rows (based on class)
for table 1}
\usage{
returnRow(x, ...)

\method{returnRow}{factor}(x, y)

\method{returnRow}{numeric}(x, y)

\method{returnRow}{character}(x, ...)

\method{returnRow}{MedIQR}(x, y)
}
\arguments{
\item{x}{the row variable}

\item{y}{the column variable}
}
\description{
These are internal functions for returning the correct rows (based on class)
for table 1
}
\section{Methods (by class)}{
\itemize{
\item \code{factor}: method for factors

\item \code{numeric}: method for numerics

\item \code{character}: method for characters - converts to factor and then
runs returnRow.factor

\item \code{MedIQR}: methods for numerics with request for Median and IQR
}}

