% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{set_global_default_analyst}
\alias{set_global_default_analyst}
\title{Set Global Default Analyst Value}
\usage{
set_global_default_analyst(analyst_name, update_template = FALSE)
}
\arguments{
\item{analyst_name}{A string containing the analyst name}

\item{update_template}{Bool whether or not to update analyst in the project template.}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the default analyst at the user level and
if possible changes the default analyst in the template for new projects.
}
\keyword{Analyst}
\keyword{options}
