% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{setAnalyst}
\alias{setAnalyst}
\title{Set Default Analyst Value}
\usage{
setAnalyst(AnalystName)
}
\arguments{
\item{AnalystName}{A string containing the analyst name}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the option CIDAtools.analyst permanently
(until you change it or reinstall CIDAtools) and will
simultanesouly change the default in New Cida Project Template.
}
\keyword{Analyst}
\keyword{options}
