% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFxns.R
\name{setPermanentAnalyst}
\alias{setPermanentAnalyst}
\title{Set Default Analyst Value}
\usage{
setPermanentAnalyst(Name)
}
\arguments{
\item{Name}{A string containing the analyst name}
}
\description{
This is an internal function that writes the Default Analyst name to the
users Rprofile.
}
