% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{SetProjectAnalyst}
\alias{SetProjectAnalyst}
\title{Set Project Analyst}
\usage{
SetProjectAnalyst(analyst_name)
}
\arguments{
\item{analyst_name}{A string containing the analyst name}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the  project analyst.
This will overwrite the current value if exists.
}
