% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{SetProjectData}
\alias{SetProjectData}
\title{Set data for project}
\usage{
SetProjectData(Parameter, Value)
}
\arguments{
\item{Parameter}{Project Parameter to be set}

\item{Value}{Value to set to project parameter}
}
\description{
Allows you to set misc project data parameters
for Project Name, Analyst, or PI recommend you use specific function
}
