% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_refactored_message.R
\name{SetProjectData}
\alias{SetProjectData}
\title{Set data for project}
\usage{
SetProjectData(Parameter, Value)
}
\description{
Allows you to set misc project data parameters
for Project Name, Analyst, or PI recommend you use specific function
}
