% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{SetProjectName}
\alias{SetProjectName}
\title{Set Project Name}
\usage{
SetProjectName(ProjectName)
}
\arguments{
\item{ProjectName}{A string containing the analyst name}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the  project name. This will overwrite the current value if exists.
}
\keyword{ProjData}
\keyword{ProjectName}
\keyword{options}
