% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\name{SetProjectPI}
\alias{SetProjectPI}
\title{Set PI Name}
\usage{
SetProjectPI(PI)
}
\arguments{
\item{PI}{A string containing the analyst name}
}
\value{
A message stating the name has been changed.
}
\description{
This function allows you to set the Project's PI. This will overwrite the
current value if exists.
}
\keyword{PI}
\keyword{ProjData}
\keyword{options}
