% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fns.R
\name{tidy.power.htest}
\alias{tidy.power.htest}
\title{Helper for pwr package version of power fns. - To be deprecated}
\usage{
\method{tidy}{power.htest}(x, ...)
}
\description{
Helper for pwr package version of power fns. - To be deprecated
}
