% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fns.R
\name{vec_power}
\alias{vec_power}
\title{Vectorized power estimates - To be deprecated}
\usage{
vec_power(fun = stats::power.t.test, ...)
}
\arguments{
\item{fun}{a power calculating function; <function>}

\item{...}{the arguments for the power calculating function assigned to `fun`;}
}
\value{
tibble of results
}
\description{
This function allows you to use power.t.test, power.prop.test, etc in
vectorized fashion and return a table of results
}
\examples{
# single non-vectorized output
vec_power(fun = power.t.test, n = 100, delta = 1, sd = 1, sig.level = 0.05)

# multiple vectorized output
vec_power(fun = power.t.test, n = 80:100, delta = 1, sd = 1, sig.level = 0.05)

# every combination of arguments vectorized output
vec_power(fun = power.t.test, n = 90:100, delta = 1, sd = seq(0, 1, length=10), sig.level = 0.05)

}
